Auto sandboxing: fixed bug with fp memory operations
x87 instructions such as frstor and fsave were claiming to
modify %rsp if %rsp was the base register in the memory
operand. This was changed so that if an instruction claims to
define a register, we check that the first operand is
actually a register, to eliminate the superfluous stack
fixup.

R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/1275703005 .