Fix llvm-dis to not crash on pnacl bitcode files with errors.
The previous code would crash when the pnacl bitcode file was
erroneous. The example bad-dis-llvm.tbc contains an example where
llvm-dis crashed. The crash was due to the code trying to print out a
corrupted module.

The fix was to terminate llvm-dis if an error occurred during parsing.

BUG=None
R=jfb@chromium.org

Review URL: https://codereview.chromium.org/1286303002 .