Fix handling errors in switch instruction by PNaCl bitcode reader.
The bitcode reader for the switch instruction has a couple of errors,
which can cause asserts/crashes to occur. One problem was that it
didn't check if the branch block of a case was defined. The other
problem was that if an error occurred after the switch instruction was
created, an assertion was fired for not deleting the switch
instruction.

BUG=None
R=jfb@chromium.org

Review URL: https://codereview.chromium.org/1283353002 .