Fix PNaCl bitcode reader's handling of basic block vector.
The bitcode reader uses a vector to hold the sequence of blocks within
a function. Previously, on bitcode record FUNC_CODE_DECLAREBLOCKS, it
would resize the vector based on the bitcode parameter. However, it
does not check if this number makes sense. For very large numbers, it
would cause the corresponding executable to hang while it tried to
meet an unreasable memory request.

This patch fixes this by not allowing the size to be any larger than
the number of bits defining the function (in the corresponding
bitstream).

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4291
R=jfb@chromium.org

Review URL: https://codereview.chromium.org/1290793002 .