PNaCl dynamic linking: Add initial ConvertToPSO pass, handling exports only
This is an initial version of a ConvertToPSO pass.  It builds arrays
for exported symbols' names and values, but no hash table yet, so
looking up an exported symbol currently requires a O(n)-time search.

The pass doesn't handle imported symbols yet.  I'll implement that in
a follow-on change.

BUG=https://bugs.chromium.org/p/nativeclient/issues/detail?id=4351
TEST=run_pnacl_dynamic_loading_test_irt, in separate NaCl-side change
R=phosek@chromium.org

Review URL: https://codereview.chromium.org/1523323006 .