Fix PNaCl bitcode writer to always add types of elided pointers.
The problem was with the NaClValueEnumerator, and the way it collects
types (used in the module). Previously, if a pointer cast instruction
was elided, it didn't add the type of the elided replacement for the
pointer cast.

The rationale for this was that the elided replacement was one of the
values where types were already collected (instructions, parameters,
globals, etc.). However, this wasn't quite correct. If one casted
a literal (i.e. constant) to a pointer type, no type signature
was collected for the literal.

This code corrects this problem.

BUG= https://bugs.chromium.org/p/nativeclient/issues/detail?id=4353
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/1614493002 .