Auto-sandboxing: ignore instructions with no register memory operands
This is a generalization of the dmb part of ra7e65db5.
Some instructions have the mayLoad/mayStore bits but no memory operands,
e.g. DMB, or have expression operands (e.g. LDR with a label operand or
ADR). If there are no memory operands, or the memory operand is not a reg,
don't modify the instruction.

R=phosek@chromium.org, jfb@chromium.org

Review URL: https://codereview.chromium.org/1643023004 .