Auto-sandboxing: use ADDrr instead of ADDrsi when emitting ARM sandbox instrs
An ADDrsi op with a shift type of no_shift is equivalent to an ADDrr op;
the asm printer does not care about the distinction and prints it correctly
but the object file encoder asserts that the shift type of ADDrsi isn't
no_shift. So create ADDrr when applicable.

R=phosek@chromium.org

Review URL: https://codereview.chromium.org/1654333002 .