Add a tool to add names for globals and functions
ABI conforming pexes have no names for globals/functions.  This leads to
llc and subzero differing in the names they assign to symbols, which makes
A/B builds crash.  This tool takes an ABI conforming pexe and adds symbol names
to produce a new pexe, which can be used with A/B builds.

BUG=
R=dschuff@chromium.org, jpp@chromium.org, stichnot@chromium.org

Review URL: https://codereview.chromium.org/1665073003 .