PNaCl dynamic linking: Added a Hash Table to quickly access exported symbols.
Inspired by the GNU ELF's ".hash" section, this CL implements
a simplified version of that hash table (for example, with no
bloom filter).

This CL also requires sorting the list of exported symbols with
respect to the hashing function. Details are included in the
commented documentation.

This CL is tested by:
https://codereview.chromium.org/1710563002/

BUG=https://bugs.chromium.org/p/nativeclient/issues/detail?id=4351
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/1705263002 .