PNaCl: ExpandTls: Generate byte offsets instead of struct field indexes
The original version of the ExpandTls pass generates a struct type
that represents the thread-local variables block.  It calculates the
address of a thread-local variable by casting the thread pointer to
that struct type and using getelementptr to calculate the addresses of
struct fields.  That generates relatively readable code and it
provides some type-checking, but it's rather complicated, especially
because we had to have two struct types (for initialized and BSS
variables).

We can simplify the code by generating byte offsets instead.  We were
already calculating the byte offsets anyway in order to add alignment
padding.

This will make it easier to reuse this code for the dynamic linking
case.

Use uint32_t for offsets for consistency and because having a >4GB TLS
area is not very reasonable.

Testing:
 * expand-tls-aligned.ll: Add tests for the generated offsets, similar
   to what expand-tls.ll tests for.
 * Remove expand-tls-bss.ll because expand-tls-aligned.ll already
   tests handling of BSS variables.

BUG=https://bugs.chromium.org/p/nativeclient/issues/detail?id=4351
TEST=test/Transforms/NaCl/expand-tls*.ll
R=smklein@chromium.org

Review URL: https://codereview.chromium.org/1765193002 .