Correct error reporting bit for parallel PNaCl bitcode parses.
This CL adds a way to deal with the fact that parallel parses of
function blocks (in PNaCl bitcode parsers) parse an extract buffer of
bits surrounding the function block.

It does this by adding the field ErrorOffset to a bitstream reader,
and the bitstream cursor function getErrorBitNo() to convert bit
positions in the bitstream reader by adding the number of bytes
specified by field ErrorOffset.

BUG= https://bugs.chromium.org/p/nativeclient/issues/detail?id=4363
R=stichnot@chromium.org

Review URL: https://codereview.chromium.org/1851163002 .