Create one-off tool pnacl-hack-memset
This tool takes a pexe, looks for calls to memset that (were) broken in
subzero, and does a patch to such calls.

That is, if the 2nd argument is a negative i8 (i.e. 0x80 through 0xff),
and the third argument is a constant, the code adds zero to the third
argument. This fix disables the (formerly) broken subzero optimization.

This hack is only necessary until about 10/31/16, when M54 gets released
for chromeos.

Note: Has command-line option --show-fixes if you want to see the calls
that were fixed.

BUG=None
R=stichnot@chromium.org

Review URL: https://codereview.chromium.org/2262373002 .