Cherry-pick r237153: [Mips] Return false for isFPCloseToIncomingSP()
On Mips, frame pointer points to the same side of the frame as the stack
pointer. This function is used to decide where to put register scavenging
spill slot. So far, it was put on the wrong side of the frame, and thus it
was too far away from $fp when frame was larger than 2^15 bytes.

Patch by Vladimir Radosavljevic.

http://reviews.llvm.org/D8895

BUG= pnacl-llc aborts on run_vector_extension_test for MIPS
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/2470173003 .